# CBG-4147: Improve Audit performance (JSON marshalling / expandFields)