CBG-4147

## benchstat before/after

```
$ benchstat base-stdlib.txt cbg-4147.txt
goos: darwin
goarch: arm64
pkg: github.com/couchbase/sync_gateway/base
                                    │ base-stdlib.txt │            cbg-4147.txt             │
                                    │     sec/op      │   sec/op     vs base                │
AuditFieldwork/additional_fields-10       5.370µ ± 2%   2.456µ ± 2%  -54.26% (p=0.000 n=10)

                                    │ base-stdlib.txt │             cbg-4147.txt             │
                                    │      B/op       │     B/op      vs base                │
AuditFieldwork/additional_fields-10      5.422Ki ± 0%   2.117Ki ± 0%  -60.96% (p=0.000 n=10)

                                    │ base-stdlib.txt │            cbg-4147.txt            │
                                    │    allocs/op    │ allocs/op   vs base                │
AuditFieldwork/additional_fields-10        71.00 ± 0%   24.00 ± 0%  -66.20% (p=0.000 n=10)
```

## Changes (ordered by improvement)
- Swap to `jsoniter` for marshalling audit event fields **(~2x speedup)**
- Build common nested audit objects as `json.RawMessage` manually to avoid marshal overhead
- Pre-allocate audit fields for marshalling at correct size (elimiates map growth - reduces `expandFields` to 1 alloc)
- Optimise `merge` method to avoid map growth/re-allocs
- Use `MarshalToString` and avoid `[]byte`->`string` conversion
- Conditional `Printf`/`Print` for FileLogger based on `len(args)`

## Profiling (with benchmark enhancements on `main`)

`base.Audit`

![Screenshot 2024-08-12 at 17 34 35](https://github.com/user-attachments/assets/3cf69f2b-5656-47ad-a096-582ed6eebdee)

`expandFields`

![Screenshot 2024-08-12 at 17 35 46](https://github.com/user-attachments/assets/bf30a448-3736-4b14-a583-a13616ef809d)



## [Integration Tests](https://jenkins.sgwdev.com/job/SyncGateway-Integration/build?delay=0sec)
- [x] `GSI=true,xattrs=true` https://jenkins.sgwdev.com/job/SyncGateway-Integration/2646/